transcriptions migration, muser password validation fix
