+++ Scriber/backend/backend/urls.py
@@ -20,6 +20,7 @@ from scriber import views
 
 router = routers.DefaultRouter()
 router.register(r'users', views.UserViewSet)
+router.register(r'transcriptions', views.TranscriptionViewSet)
 
 urlpatterns = [
     url(r'^', include(router.urls)),
