+++ Scriber/backend/scriber/views.py
@@ -3,6 +3,8 @@ from rest_framework import permissions, routers, viewsets, authentication
 from django.contrib.auth.models import User
 from oauth2_provider.contrib.rest_framework import TokenHasReadWriteScope, TokenHasScope
 from scriber.serializers import UserSerializer
+from scriber.models import Transcription
+from scriber.serializers import TranscriptionSerializer
 # Create your views here.
 
 class UserViewSet(viewsets.ModelViewSet):
@@ -11,3 +13,7 @@ class UserViewSet(viewsets.ModelViewSet):
     authentication_classes = (authentication.TokenAuthentication,)
     queryset = User.objects.all()
     serializer_class = UserSerializer
+
+class TranscriptionViewSet(viewsets.ModelViewSet):
+    queryset = Transcription.objects.all()
+    serializer_class = TranscriptionSerializer
